# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manypost API",
    "version": "0.0.1",
    "description": "API do manypost — agendamento/publicação multicanal. Humanos usam sessão Clerk; máquinas usam `Authorization: Bearer mp_live_…`. Erros seguem problem+json (RFC 9457): o campo `title` carrega o código estável do erro."
  },
  "servers": [
    {
      "url": "/",
      "description": "mesma origem (proxy do Next)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "sessão Clerk humana — Authorization: Bearer <token>"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key mp_live_ para superfícies de máquina — Authorization: Bearer <chave>"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "código estável do erro",
            "example": "common.not_found"
          },
          "status": {
            "type": "integer",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "post não encontrado"
          },
          "extra": {
            "description": "campos extras (ex.: issues de validação)"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "ContentOverview": {
        "type": "object",
        "properties": {
          "brand": {
            "$ref": "#/components/schemas/ContentBrand"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ContentCapabilities"
          },
          "counts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "foundationFilled": {
            "type": "integer"
          },
          "foundationTotal": {
            "type": "integer"
          }
        },
        "required": [
          "brand",
          "capabilities",
          "counts",
          "foundationFilled",
          "foundationTotal"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/content-machine/overview": {
      "get": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Estado da máquina: identidade, capacidades e contagem por etapa",
        "description": "Na primeira chamada, configura a organização (prompts, fórmulas e fundação vazia).",
        "responses": {
          "200": {
            "description": "visão geral",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentOverview"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content-machine/brand": {
      "put": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Atualiza a identidade visual e as preferências de publicação",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "logoMediaId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "logoDarkMediaId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "palette": {
                    "$ref": "#/components/schemas/ContentPalette"
                  },
                  "slogan": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "signature": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "tone": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "defaultChannelId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "ctaChannel": {
                    "type": "string",
                    "enum": [
                      "direct",
                      "whatsapp"
                    ]
                  },
                  "whatsappNumber": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "publishHour": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 23
                  },
                  "timezone": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "autoApprove": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "identidade",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentBrand"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content-machine/brand/palette": {
      "post": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Extrai a paleta de cores da logo",
        "description": "Sem `mediaId`, usa a logo principal. Sem logo definida, a imagem enviada vira a logo.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediaId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "identidade com a paleta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentBrand"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "capacidade não disponível nesta instalação",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "provedor externo falhou — o pedido pode ser repetido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content-machine/foundation": {
      "get": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Documentos da fundação editorial",
        "responses": {
          "200": {
            "description": "fundação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentFoundation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content-machine/foundation/{key}": {
      "put": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Grava um documento da fundação",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "produtos",
                "icp",
                "personagem",
                "marca",
                "escada",
                "pilares"
              ]
            },
            "required": true,
            "name": "key",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "type": "string",
                    "maxLength": 30000
                  },
                  "validUntil": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  }
                },
                "required": [
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "documento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentFoundation"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content-machine/prompts": {
      "get": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Prompts da máquina, com o histórico de versões",
        "responses": {
          "200": {
            "description": "prompts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentPrompt"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/content-machine/hooks": {
      "get": {
        "tags": [
          "content-machine"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Fórmulas de gancho que a pauta usa",
        "responses": {
          "200": {
            "description": "fórmulas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentHook"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}